refactor(terminal): replace clipboard binary protocol with S3 URL paste for image delivery
Replace the complex clipboard shim infrastructure (binary protocol → relay →
runner → native clipboard → Ctrl+V injection) with a simple upload-to-S3 +
paste-URL-as-text approach. Images pasted in the terminal are now uploaded via
the existing file upload API and the presigned URL is sent as standard terminal
input text.

Changes:
- Add shared uploadImage() utility (web/src/lib/api/file.ts)
- Terminal paste handler now uploads image and sends URL as text input
- Remove MsgType.ImagePaste from relay and runner binary protocols
- Remove runner/internal/clipboard package entirely (shim + native backends)
- Remove SetImagePasteHandler from relay client interface
- Remove clipboard setup from PodBuilder and Pod struct
- Update block-editor to use shared upload utility
