+++ runner/internal/relay/mock_client.go
@@ -36,7 +36,6 @@ type MockClient struct {
 	onInput        InputHandler
 	onResize       ResizeHandler
 	onClose        CloseHandler
-	onImagePaste   ImagePasteHandler
 	onReconnect    func()
 	onTokenExpired func() string
 }
@@ -134,13 +133,6 @@ func (m *MockClient) SetCloseHandler(handler CloseHandler) {
 	m.mu.Lock()
 	defer m.mu.Unlock()
 	m.onClose = handler
-}
-
-// SetImagePasteHandler implements RelayClient.
-func (m *MockClient) SetImagePasteHandler(handler ImagePasteHandler) {
-	m.mu.Lock()
-	defer m.mu.Unlock()
-	m.onImagePaste = handler
 }
 
 // SetReconnectHandler implements RelayClient.
