+++ runner/internal/runner/pod_builder.go
@@ -3,7 +3,6 @@ package runner
 import (
 	runnerv1 "github.com/anthropics/agentsmesh/proto/gen/go/runner/v1"
 	"github.com/anthropics/agentsmesh/runner/internal/client"
-	"github.com/anthropics/agentsmesh/runner/internal/clipboard"
 	"github.com/anthropics/agentsmesh/runner/internal/config"
 	"github.com/anthropics/agentsmesh/runner/internal/terminal/vt"
 	"github.com/anthropics/agentsmesh/runner/internal/workspace"
@@ -22,10 +21,6 @@ type PodBuilderDeps struct {
 	// ProgressSender sends pod initialization progress.
 	// Can be nil if progress reporting is not needed.
 	ProgressSender client.ProgressSender
-
-	// Clipboard provides clipboard backend for image paste support.
-	// Can be nil; if nil, clipboard support is disabled.
-	Clipboard clipboard.Backend
 }
 
 // PodBuilder builds pods using the Builder pattern.
@@ -76,7 +71,6 @@ func NewPodBuilderFromRunner(runner *Runner) *PodBuilder {
 	deps := PodBuilderDeps{
 		Config:         runner.cfg,
 		ProgressSender: runner.conn,
-		Clipboard:      runner.clipboard,
 	}
 	// Explicitly set Workspace only if not nil to avoid interface nil comparison issues
 	if runner.workspace != nil {
